refactor and add integration tests
