+++ src/dex/index.ts
@@ -69,7 +69,6 @@ import { SwaapV2 } from './swaap-v2/swaap-v2';
 import { SpiritSwapV3 } from './quickswap/spiritswap-v3';
 import { TraderJoeV21 } from './trader-joe-v2.1';
 import { PancakeswapV3 } from './pancakeswap-v3/pancakeswap-v3';
-// import { RamsesV2 } from './ramses-v2/ramses-v2';
 import { Algebra } from './algebra/algebra';
 
 const LegacyDexes = [
@@ -138,7 +137,6 @@ const Dexes = [
   MaverickV1,
   Camelot,
   SwaapV2,
-  // RamsesV2,
 ];
 
 export type LegacyDexConstructor = new (dexHelper: IDexHelper) => IDexTxBuilder<
