--- src/dex/ramses-v2/contract-math/FixedPoint128.ts
@@ -1,3 +1,0 @@
-export class FixedPoint128 {
-  static readonly Q128 = 0x100000000000000000000000000000000n;
-}
