--- src/dex/ramses-v2/contract-math/FixedPoint96.ts
@@ -1,4 +1,0 @@
-export class FixedPoint96 {
-  static readonly RESOLUTION = 96n;
-  static readonly Q96 = 0x1000000000000000000000000n;
-}
