--- src/dex/ramses-v2/contract-math/UnsafeMath.ts
@@ -1,5 +1,0 @@
-export class UnsafeMath {
-  static divRoundingUp(x: bigint, y: bigint) {
-    return (x + y - 1n) / y;
-  }
-}
