--- src/dex/ramses-v2/contract-math/utils.ts
@@ -1,11 +1,0 @@
-export function _mulmod(x: bigint, y: bigint, m: bigint): bigint {
-  return m === 0n ? 0n : (x * y) % m;
-}
-
-export function _lt(x: bigint, y: bigint) {
-  return x < y ? 1n : 0n;
-}
-
-export function _gt(x: bigint, y: bigint) {
-  return x > y ? 1n : 0n;
-}
