+++ src/dex/uniswap-v3/config.ts
@@ -111,8 +111,7 @@ export const UniswapV3Config: DexConfigMap<DexParams> = {
     [Network.ARBITRUM]: {
       factory: '0xAA2cd7477c451E703f3B9Ba5663334914763edF8',
       deployer: '0xb3e423ab9cE6C03D98326A3A2a0D7D96b0829f22',
-      quoter: '0x61fFE014bA17989E743c5F6cB21bF9697530B21e',
-      // quoter: '0xAA20EFF7ad2F523590dE6c04918DaAE0904E3b20',
+      quoter: '0xAA20EFF7ad2F523590dE6c04918DaAE0904E3b20',
       router: '0xAA23611badAFB62D37E7295A682D21960ac85A90',
       supportedFees: SUPPORTED_FEES,
       stateMulticall: '0x50EE4112Cab9c79812F23bE079aB3911395ACc8e',
