+++ src/dex/uniswap-v3/uniswap-v3.ts
@@ -42,7 +42,6 @@ import {
 import { UniswapV3Config, Adapters, PoolsToPreload } from './config';
 import { UniswapV3EventPool } from './uniswap-v3-pool';
 import UniswapV3RouterABI from '../../abi/uniswap-v3/UniswapV3Router.abi.json';
-import UniswapV3QuoterABI from '../../abi/uniswap-v3/UniswapV3Quoter.abi.json';
 import UniswapV3QuoterV2ABI from '../../abi/uniswap-v3/UniswapV3QuoterV2.abi.json';
 import UniswapV3MultiABI from '../../abi/uniswap-v3/UniswapMulti.abi.json';
 import DirectSwapABI from '../../abi/DirectSwap.json';
@@ -370,7 +369,6 @@ export class UniswapV3
     side: SwapSide,
     pools: UniswapV3EventPool[],
   ): Promise<ExchangePrices<UniswapV3Data> | null> {
-    console.log('POOLS: ', pools);
     if (pools.length === 0) {
       return null;
     }
@@ -424,7 +422,6 @@ export class UniswapV3
       ),
     );
 
-    console.log('QUOTER: ', this.config.quoter);
     const calldata = pools.map(pool =>
       _amounts.map(_amount => ({
         target: this.config.quoter,
@@ -580,18 +577,17 @@ export class UniswapV3
 
       if (selectedPools.length === 0) return null;
 
-
       const poolsToUse = selectedPools.reduce(
         (acc, pool) => {
           let state = pool.getState(blockNumber);
-          // if (state === null || pool.poolAddress === '0x6059cf1c818979bccac5d1f015e1b322d154592f') {
+          if (state === null) {
             this.logger.trace(
               `${this.dexKey}: State === null. Fallback to rpc ${pool.name}`,
             );
             acc.poolWithoutState.push(pool);
-          // } else {
-          //   acc.poolWithState.push(pool);
-          // }
+          } else {
+            acc.poolWithState.push(pool);
+          }
           return acc;
         },
         {
@@ -599,7 +595,6 @@ export class UniswapV3
           poolWithoutState: [] as UniswapV3EventPool[],
         },
       );
-
 
       const rpcResultsPromise = this.getPricingFromRpc(
         _srcToken,
@@ -699,9 +694,6 @@ export class UniswapV3
         }),
       );
       const rpcResults = await rpcResultsPromise;
-
-
-      console.log('RPC RESULTS: ', rpcResults);
 
       const notNullResult = result.filter(
         res => res !== null,
